\name{adjPValuesPlot}
\alias{adjPValuesPlot}
\title{A function plotting adjusted p-values...}
\usage{adjPValuesPlot(adjPValues, alpha)}
\description{A function plotting adjusted p-values}
\author{MuToss-Coding Team}
\arguments{
\item{adjPValues}{A numeric containing the adjusted pValues to plot.}
\item{alpha}{A numeric containing the overall type I error rate alpha..}
}
