\name{fisher22.marginal}
\alias{fisher22.marginal}
\title{Fisher (2x2) table association analysis for calculating all marginal p-values...}
\usage{fisher22.marginal(data, model)}
\description{Fisher (2x2) table association analysis for calculating all marginal p-values}
\value{A list containing the m marginal p-values}
\author{ThorstenDickhaus}
\arguments{\item{data}{A tensor of dimension (2x2xm) where m is the number of endpoints (genes, etc.)}
\item{model}{A model object indicating that this type of analysis shall be performed}}
