\name{fisher23.marginal}
\alias{fisher23.marginal}
\title{Fisher-type (2x3) table association analysis for calculating all marginal p-values...}
\usage{fisher23.marginal(data, model)fisher23.marginal(data, model)
fisher23.marginal(data, model)
fisher23.marginal(data, model)
fisher23.marginal(data, model)
}
\description{Fisher-type (2x3) table association analysis for calculating all marginal p-values}
\value{A list containing the m marginal p-values\code{fisher23.marginal}: A list containing the m marginal p-values


\code{fisher23.marginal}: A list containing the m marginal p-values


\code{fisher23.marginal}: A list containing the m marginal p-values


\code{fisher23.marginal}: A list containing the m marginal p-values


}
\author{ThorstenDickhaus}
\arguments{\item{data}{A tensor of dimension (2x3xm) where m is the number of endpoints (genes, etc.)}
\item{model}{A model object indicating that this type of analysis shall be performed}}
\alias{fisher23.marginal}

\alias{fisher23.marginal}

\alias{fisher23.marginal}

\alias{fisher23.marginal}

