\name{fisher23.model}
\alias{fisher23.model}
\title{Fisher-type (2x3) table as model for marginal hypotheses testing problems...}
\description{Fisher-type (2x3) table as model for marginal hypotheses testing problems}
\value{A list containing the model description\code{fisher23.model}: A list containing the model description


\code{fisher23.model}: A list containing the model description


\code{fisher23.model}: A list containing the model description


\code{fisher23.model}: A list containing the model description


}
\author{ThorstenDickhaus}
\alias{fisher23.model}

\alias{fisher23.model}

\alias{fisher23.model}

\alias{fisher23.model}

