\name{ftest.marginal}
\alias{ftest.marginal}
\title{ftest.marginal}
\usage{ftest.marginal(data, model, robust)ftest.marginal(data, model, robust)
ftest.marginal(data, model, robust)
ftest.marginal(data, model, robust)
ftest.marginal(data, model, robust)
}
\alias{ftest.marginal}

\alias{ftest.marginal}

\alias{ftest.marginal}

\alias{ftest.marginal}

