\name{gatherParameters}
\alias{gatherParameters}
\title{Extracts the parameters from the MutossSim objects in listOfObjects and...}
\usage{gatherParameters(listOfObjects)gatherParameters(listOfObjects)
gatherParameters(listOfObjects)
gatherParameters(listOfObjects)
gatherParameters(listOfObjects)
}
\description{Extracts the parameters from the MutossSim objects in listOfObjects and
creates a data.frame from this. Every used parameter gets its own column.
Every row corresponds to one MutossSim object. If a MutossSim object A does not
contain a parameter P that another object B does then the row for object A
will have "" in the column for the parameter P.}
\value{A data.frame with rows for every MutossSim object and columns for
the used parameter.\code{gatherParameters}: A data.frame with rows for every MutossSim object and columns for
the used parameter.


\code{gatherParameters}: A data.frame with rows for every MutossSim object and columns for
the used parameter.


\code{gatherParameters}: A data.frame with rows for every MutossSim object and columns for
the used parameter.


\code{gatherParameters}: A data.frame with rows for every MutossSim object and columns for
the used parameter.


}
\author{MarselScheer}
\arguments{\item{listOfObjects}{A list with MutossSim objects}

}
\alias{gatherParameters}

\alias{gatherParameters}

\alias{gatherParameters}

\alias{gatherParameters}

