\name{linearStepUp}
\alias{linearStepUp}
\title{Linear Step Up Service Function...}
\usage{linearStepUp(sorted, q, m, adjust=FALSE, m0=m, pi0, constant=1)linearStepUp(sorted, q, m, adjust=FALSE, m0=m, pi0, constant=1)
linearStepUp(sorted, q, m, adjust=FALSE, m0=m, pi0, constant=1)
linearStepUp(sorted, q, m, adjust=FALSE, m0=m, pi0, constant=1)
linearStepUp(sorted, q, m, adjust=FALSE, m0=m, pi0, constant=1)
}
\description{Linear Step Up Service Function}
\details{A Mutoss service function called by other procedures.\code{linearStepUp}: A Mutoss service function called by other procedures.


\code{linearStepUp}: A Mutoss service function called by other procedures.


\code{linearStepUp}: A Mutoss service function called by other procedures.


\code{linearStepUp}: A Mutoss service function called by other procedures.


}
\value{A list containing the following objects: 
\item{Cutoff}{The largest p-value of rejected hypotheses.}
\item{Pvals}{A data frame containing the original p-values, critical values, adjusted p-values and rejections.}\code{linearStepUp}: A list containing the following objects: 
\item{Cutoff}{The largest p-value of rejected hypotheses.}
\item{Pvals}{A data frame containing the original p-values, critical values, adjusted p-values and rejections.}


\code{linearStepUp}: A list containing the following objects: 
\item{Cutoff}{The largest p-value of rejected hypotheses.}
\item{Pvals}{A data frame containing the original p-values, critical values, adjusted p-values and rejections.}


\code{linearStepUp}: A list containing the following objects: 
\item{Cutoff}{The largest p-value of rejected hypotheses.}
\item{Pvals}{A data frame containing the original p-values, critical values, adjusted p-values and rejections.}


\code{linearStepUp}: A list containing the following objects: 
\item{Cutoff}{The largest p-value of rejected hypotheses.}
\item{Pvals}{A data frame containing the original p-values, critical values, adjusted p-values and rejections.}


}
\author{JonathanRosenblatt}
\arguments{\item{sorted}{Numeric vector of sorted pvalues}
\item{q}{Error rate to control for.}
\item{m}{Number of hypothesis tested.}
\item{adjust}{Logical value for p-value adjustmet (unusable).}
\item{m0}{Known or estimated number of true null hypotheses.}
\item{pi0}{Known or estimated proportion of true null hypothesis. Is redundant when \code{m0} is specified.}
\item{constant}{}}
\alias{linearStepUp}

\alias{linearStepUp}

\alias{linearStepUp}

\alias{linearStepUp}

