\name{mu.test.class}
\alias{mu.test.class}
\title{Test for compatible classes for comparison...}
\usage{mu.test.class(classes)mu.test.class(classes)
mu.test.class(classes)
mu.test.class(classes)
mu.test.class(classes)
}
\description{Test for compatible classes for comparison}
\details{Internal function of muToss package.
Takes list of classes and tests if they are all muToss.\code{mu.test.class}: Internal function of muToss package.
Takes list of classes and tests if they are all muToss.


\code{mu.test.class}: Internal function of muToss package.
Takes list of classes and tests if they are all muToss.


\code{mu.test.class}: Internal function of muToss package.
Takes list of classes and tests if they are all muToss.


\code{mu.test.class}: Internal function of muToss package.
Takes list of classes and tests if they are all muToss.


}
\value{True if all classes are muToss. False otherwise.\code{mu.test.class}: True if all classes are muToss. False otherwise.


\code{mu.test.class}: True if all classes are muToss. False otherwise.


\code{mu.test.class}: True if all classes are muToss. False otherwise.


\code{mu.test.class}: True if all classes are muToss. False otherwise.


}
\author{MuToss-Coding Team.}
\arguments{\item{classes}{A list of class names.}

}
\alias{mu.test.class}

\alias{mu.test.class}

\alias{mu.test.class}

\alias{mu.test.class}

