\name{mu.test.name}
\alias{mu.test.name}
\title{Tests if all hypotheses have the same names.}
\usage{mu.test.name(hyp.names)}
\description{Tests if all hypotheses have the same names.}
\details{Internal muToss function.}
\value{Gives a notice when hypotheses names in different procedures are found different.}
\author{MuToss-Coding Team}
\arguments{\item{hyp.names}{Character vector of hypotheses names.}}
