\name{mu.test.same.data}
\alias{mu.test.same.data}
\title{Tests if the same pvalues were used by different procedures.}
\usage{mu.test.same.data(pvals)mu.test.same.data(pvals)
mu.test.same.data(pvals)
mu.test.same.data(pvals)
mu.test.same.data(pvals)
}
\description{Tests if the same pvalues were used by different procedures.}
\details{Internal muToss function.\code{mu.test.same.data}: Internal muToss function.


\code{mu.test.same.data}: Internal muToss function.


\code{mu.test.same.data}: Internal muToss function.


\code{mu.test.same.data}: Internal muToss function.


}
\value{Stops if different data was used by different procedures.\code{mu.test.same.data}: Stops if different data was used by different procedures.


\code{mu.test.same.data}: Stops if different data was used by different procedures.


\code{mu.test.same.data}: Stops if different data was used by different procedures.


\code{mu.test.same.data}: Stops if different data was used by different procedures.


}
\author{MuToss-Coding Team}
\arguments{\item{pvals}{Data frame of p-values used by each procedure.}

}
\alias{mu.test.same.data}

\alias{mu.test.same.data}

\alias{mu.test.same.data}

\alias{mu.test.same.data}

