\name{multiple.down.adjust}
\alias{multiple.down.adjust}
\title{A service function used by multiple...}
\usage{multiple.down.adjust(sorted, m)}
\description{A service function used by \code{multiple.down}}
\author{JonathanRosenblatt}
\arguments{
\item{sorted}{A sorted pvalue vector.}
\item{m}{The number of hypotheses tested.}
}
