\name{mutoss.fisher22.model}
\alias{mutoss.fisher22.model}
\title{Fisher (2x2) table as model for marginal hypotheses testing problems...}
\description{Fisher (2x2) table as model for marginal hypotheses testing problems}
\value{A list containing the model description\code{mutoss.fisher22.model}: A list containing the model description


\code{mutoss.fisher22.model}: A list containing the model description


\code{mutoss.fisher22.model}: A list containing the model description


\code{mutoss.fisher22.model}: A list containing the model description


}
\author{ThorstenDickhaus}
\alias{mutoss.fisher22.model}

\alias{mutoss.fisher22.model}

\alias{mutoss.fisher22.model}

\alias{mutoss.fisher22.model}

