\name{mutoss.models}
\alias{fisher23.model}
\alias{fisher22.model}
\alias{twosamp.model}
\alias{paired.model}
\alias{onesamp.model}
\alias{ftest.model}
\title{Mutoss Models}
\description{Fisher-type (2x3) table as model for marginal hypotheses testing problems,
Fisher (2x2) table as model for marginal hypotheses testing problems and others...}
\value{A list containing the model description}