\name{pValuesPlot}
\alias{pValuesPlot}
\title{A function plotting p-values...}
\usage{pValuesPlot(pValues)pValuesPlot(pValues)
pValuesPlot(pValues)
pValuesPlot(pValues)
pValuesPlot(pValues)
}
\description{A function plotting p-values}
\author{MarselScheer}
\arguments{\item{pValues}{}

}
\alias{pValuesPlot}

\alias{pValuesPlot}

\alias{pValuesPlot}

\alias{pValuesPlot}

