\name{weightMatrix}
\alias{weightMatrix}
\title{weightMatrix}
\usage{weightMatrix(n, type=c("UserDefined", "Tukey", "AVE", "Dunnett",
    "Sequen", "Changepoint", "Marcus", "McDermott", "Williams",
    "UmbrellaWilliams"), base=1, contrast.matrix=NULL)weightMatrix(n, type=c("UserDefined", "Tukey", "AVE", "Dunnett", "Sequen",
    "Changepoint", "Marcus", "McDermott", "Williams",
    "UmbrellaWilliams"), base=1, contrast.matrix=NULL)
weightMatrix(n, type=c("UserDefined", "Tukey", "AVE", "Dunnett", "Sequen",
    "Changepoint", "Marcus", "McDermott", "Williams",
    "UmbrellaWilliams"), base=1, contrast.matrix=NULL)
weightMatrix(n, type=c("UserDefined", "Tukey", "AVE", "Dunnett", "Sequen",
    "Changepoint", "Marcus", "McDermott", "Williams",
    "UmbrellaWilliams"), base=1, contrast.matrix=NULL)
weightMatrix(n, type=c("UserDefined", "Tukey", "AVE", "Dunnett", "Sequen",
    "Changepoint", "Marcus", "McDermott", "Williams",
    "UmbrellaWilliams"), base=1, contrast.matrix=NULL)
}
\alias{weightMatrix}

\alias{weightMatrix}

\alias{weightMatrix}

\alias{weightMatrix}

