\name{adjPValuesPlot}
\alias{adjPValuesPlot}
\title{A function plotting adjusted p-values...}
\usage{adjPValuesPlot(adjPValues, alpha)adjPValuesPlot(adjPValues, alpha)
adjPValuesPlot(adjPValues, alpha)
adjPValuesPlot(adjPValues, alpha)
adjPValuesPlot(adjPValues, alpha)
}
\description{A function plotting adjusted p-values}
\author{Kornel}
\arguments{\item{pValues}{}

}
\alias{adjPValuesPlot}

\alias{adjPValuesPlot}

\alias{adjPValuesPlot}

\alias{adjPValuesPlot}

