\name{mu.test.class}
\alias{mu.test.class}
\title{Test for compatible classes for comparison...}
\usage{mu.test.class(classes)}
\description{Test for compatible classes for comparison}
\details{Internal function of muToss package.
Takes list of classes and tests if they are all muToss.}
\value{True if all classes are muToss. False otherwise.}
\author{MuToss-Coding Team.}
\arguments{\item{classes}{A list of class names.}}
