\name{mu.test.rates}
\alias{mu.test.rates}
\title{Tests that different procedures used the same error rates.}
\usage{mu.test.rates(rates)}
\description{Tests that different procedures used the same error rates.}
\details{Internal muToss function.}
\value{A notice if error rates differ.}
\author{MuToss-Coding Team.}
\arguments{\item{rates}{Numeric vector error rates extracted fomr muToss objects.}}
