\name{mu.test.type}
\alias{mu.test.type}
\title{Tests that different procedures use the same error types.}
\usage{mu.test.type(types)mu.test.type(types)
mu.test.type(types)
mu.test.type(types)
mu.test.type(types)
}
\description{Tests that different procedures use the same error types.}
\details{Internal muToss function.\code{mu.test.type}: Internal muToss function.


\code{mu.test.type}: Internal muToss function.


\code{mu.test.type}: Internal muToss function.


\code{mu.test.type}: Internal muToss function.


}
\value{Returns a notice if error types differ.\code{mu.test.type}: Returns a notice if error types differ.


\code{mu.test.type}: Returns a notice if error types differ.


\code{mu.test.type}: Returns a notice if error types differ.


\code{mu.test.type}: Returns a notice if error types differ.


}
\author{MuToss-Coding Team.}
\arguments{\item{types}{Character vector of error types extracted from muToss objects.}

}
\alias{mu.test.type}

\alias{mu.test.type}

\alias{mu.test.type}

\alias{mu.test.type}

