\name{notterman}
\docType{data}
\alias{notterman}
\alias{notterman.grpLabel}
\alias{T.Test.tumor.vs.normal}
\title{Notterman data set}
\description{
   Notterman data set - syntactically correct Rd file - Thorsten has to send us information about the data.
}
\usage{notterman
notterman.grpLabel
T.Test.tumor.vs.normal
}
\format{
\itemize{
		\item{notterman - A data.frame containing 36 individuals with 7457 observations}
		\item{notterman.grpLabel - A vector containing 36 labels of type \code{Tumor} or \code{Normal}}
		\item{T.Test.tumor.vs.normal - A vector containing 7457 p-values}
	}
}
\source{Thorsten}
\references{
   Thorsten
}
\keyword{datasets}
